backend/requirements.txt
fastapi
uvicorn
inference
inference-sdk
supervision
opencv-python-headless
python-multipart
python-dotenv
numpy